Gateway/requirements.txt
slowapi==0.1.9
fastapi==0.115.12
httpx~=0.28.1
python-decouple~=3.8
starlette~=0.46.2
uvicorn==0.34.2